Refactoring core library usage in monorepo, make it a bit cleaner/easier to search/more standardised.
