+++ packages/backend-core/objectStore.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require("./src/objectStore"),
+  ...require("./src/objectStore/utils"),
+}
