+++ packages/backend-core/utils.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require("./src/utils"),
+  ...require("./src/hashing"),
+}
