+++ packages/server/src/api/controllers/hosting.js
@@ -1,7 +1,7 @@
 const CouchDB = require("../../db")
 const { getDeployedApps } = require("../../utilities/workerRequests")
 const { getScopedConfig } = require("@budibase/backend-core/db")
-const { Configs } = require("@budibase/backend-core").constants
+const { Configs } = require("@budibase/backend-core/constants")
 const { checkSlashesInUrl } = require("../../utilities")
 
 exports.fetchUrls = async ctx => {
