+++ packages/server/src/constants/index.js
@@ -1,6 +1,6 @@
 const { BUILTIN_ROLE_IDS } = require("@budibase/backend-core/roles")
-const { UserStatus } = require("@budibase/backend-core").constants
-const { ObjectStoreBuckets } = require("@budibase/backend-core").objectStore
+const { UserStatus } = require("@budibase/backend-core/constants")
+const { ObjectStoreBuckets } = require("@budibase/backend-core/objectStore")
 
 exports.JobQueues = {
   AUTOMATIONS: "automationQueue",
