+++ packages/worker/src/api/index.js
@@ -2,8 +2,11 @@ const Router = require("@koa/router")
 const compress = require("koa-compress")
 const zlib = require("zlib")
 const { routes } = require("./routes")
-const { buildAuthMiddleware, auditLog, buildTenancyMiddleware } =
-  require("@budibase/backend-core").auth
+const {
+  buildAuthMiddleware,
+  auditLog,
+  buildTenancyMiddleware,
+} = require("@budibase/backend-core/auth")
 
 const PUBLIC_ENDPOINTS = [
   // old deprecated endpoints kept for backwards compat
