+++ packages/worker/src/constants/index.js
@@ -1,4 +1,4 @@
-const { Configs } = require("@budibase/backend-core").constants
+const { Configs } = require("@budibase/backend-core/constants")
 
 exports.LOGO_URL =
   "https://d33wubrfki0l68.cloudfront.net/aac32159d7207b5085e74a7ef67afbb7027786c5/2b1fd/img/logo/bb-emblem.svg"
