+++ packages/worker/src/index.js
@@ -6,7 +6,7 @@ const Koa = require("koa")
 const destroyable = require("server-destroy")
 const koaBody = require("koa-body")
 const koaSession = require("koa-session")
-const { passport } = require("@budibase/backend-core").auth
+const { passport } = require("@budibase/backend-core/auth")
 const logger = require("koa-pino-logger")
 const http = require("http")
 const api = require("./api")
